requirements.txt
langchain==0.1.8
langchain-community==0.0.16
langchain-google-genai==0.0.6
google-generativeai==0.3.1
pypdf==3.17.1
sentence-transformers==2.2.2
chromadb==0.4.22
